Replace old KLV system with new
