+++ arrows/klv/klv_data_format.cxx
@@ -7,8 +7,8 @@
 
 #include "klv_data_format.h"
 
-#include "klv_0104_new.h"
-#include "klv_0601_new.h"
+#include "klv_0104.h"
+#include "klv_0601.h"
 #include "klv_1108.h"
 #include "klv_1108_metric_set.h"
 #include "klv_blob.h"
