+++ arrows/klv/klv_value.cxx
@@ -7,8 +7,8 @@
 
 #include "klv_value.h"
 
-#include "klv_0104_new.h"
-#include "klv_0601_new.h"
+#include "klv_0104.h"
+#include "klv_0601.h"
 #include "klv_1108.h"
 #include "klv_1108_metric_set.h"
 #include "klv_blob.h"
