+++ arrows/vxl/vidl_ffmpeg_video_input.cxx
@@ -11,9 +11,9 @@
 
 #include <arrows/vxl/image_container.h>
 
-#include <arrows/klv/convert_metadata.h>
+#include <arrows/klv/klv_convert_vital.h>
+#include <arrows/klv/klv_demuxer.h>
 #include <arrows/klv/misp_time.h>
-#include <arrows/klv/klv_data.h>
 
 #include <vital/types/timestamp.h>
 #include <vital/exceptions/io.h>
@@ -28,11 +28,11 @@
 
 #include <kwiversys/SystemTools.hxx>
 
-#include <mutex>
-#include <memory>
-#include <vector>
-#include <sstream>
 #include <chrono>
+#include <memory>
+#include <mutex>
+#include <sstream>
+#include <vector>
 
 namespace kwiver {
 namespace arrows {
@@ -62,8 +62,10 @@ public:
       d_have_loop_vars( false ),
       pts_of_meta_ts( 0.0 ),
       meta_ts( 0 ),
+      m_last_misp_timestamp( 0 ),
       d_frame_time( 0 ),
-      d_frame_number( 1 )
+      d_frame_number( 1 ),
+      m_klv_demuxer( m_klv_timeline )
   { }
 
   vidl_ffmpeg_istream d_video_stream;
@@ -133,6 +135,7 @@ public:
 
   double pts_of_meta_ts;            // probably seconds
   vital::time_usec_t meta_ts; // time in usec
+  uint64_t m_last_misp_timestamp;
 
   // used to create timestamp output
   vital::time_usec_t d_frame_time; // usec
@@ -140,9 +143,10 @@ public:
 
   std::string video_path; // name of video we opened
 
-  std::deque<uint8_t> md_buffer; // working buffer for metadata stream
+  std::vector< uint8_t > md_buffer; // working buffer for metadata stream
 
-  kwiver::arrows::klv::convert_metadata converter; // metadata converter object
+  klv::klv_demuxer m_klv_demuxer;
+  klv::klv_timeline m_klv_timeline;
 
   static std::mutex s_open_mutex;
 
@@ -164,68 +168,51 @@ public:
    * @return the processed metadata. If there is no metadata it returns an
    *         empty metadata vector.
    */
-  kwiver::vital::metadata_vector process_metadata( std::deque<uint8_t> const& curr_md )
+  kwiver::vital::metadata_vector process_metadata( std::vector< uint8_t > const& curr_md )
   {
-    kwiver::vital::metadata_vector retval;
-
     // Add new metadata to the end of current metadata stream
-    md_buffer.insert(md_buffer.end(), curr_md.begin(), curr_md.end());
-    kwiver::arrows::klv::klv_data klv_packet;
+    md_buffer.insert( md_buffer.end(), curr_md.begin(), curr_md.end() );
 
-    // If we have collected enough of the stream to make a KLV packet
-    while ( kwiver::arrows::klv::klv_pop_next_packet( md_buffer, klv_packet ) )
+    auto it = md_buffer.cbegin();
+    while( it != md_buffer.cend() )
     {
-      auto meta = std::make_shared<kwiver::vital::metadata>();
-
+      klv::klv_packet packet;
       try
       {
-        converter.convert( klv_packet, *(meta) );
+        packet = klv::klv_read_packet( it, std::distance( it, md_buffer.cend() ) );
       }
-      catch ( kwiver::vital::metadata_exception const& e )
+      catch( kwiver::vital::metadata_buffer_overflow const& e )
       {
-        LOG_WARN( this->d_logger, "Metadata exception: " << e.what() );
-        continue;
+        // We only have part of a packet; quit until we have more data
+        break;
       }
-
-      // If the metadata was even partially decided, then add to the list.
-      if ( ! meta->empty() )
+      catch( kwiver::vital::metadata_exception const& e )
       {
-        kwiver::vital::timestamp ts;
-        ts.set_frame( this->d_frame_number );
-
-        if ( this->d_have_frame_time )
-        {
-          ts.set_time_usec( this->d_frame_time );
-        }
-
-        meta->set_timestamp( ts );
-
-        meta->add< vital::VITAL_META_VIDEO_URI>( video_path );
-        retval.push_back( meta );
-      } // end valid metadata packet.
-    } // end while
-
-    // if no metadata from the stream, add a basic metadata item
-    // containing video name and timestamp
-    if ( retval.empty() )
-    {
-      auto meta = std::make_shared<kwiver::vital::metadata>();
-      kwiver::vital::timestamp ts;
-      ts.set_frame(this->d_frame_number);
-
-      if (this->d_have_frame_time)
-      {
-        ts.set_time_usec(this->d_frame_time);
+        LOG_ERROR( d_logger, "error while parsing KLV packet: " << e.what() );
       }
 
-      meta->set_timestamp(ts);
-
-      meta->add< vital::VITAL_META_VIDEO_URI >( video_path );
-
-      retval.push_back(meta);
+      m_klv_demuxer.demux_packet( packet );
     }
 
-    return retval;
+    // Erase the bytes we just used
+    md_buffer.erase( md_buffer.cbegin(), it );
+
+    // Get the vital metadata structure for the current frame
+    auto result =
+      klv::klv_to_vital_metadata( m_klv_timeline, m_last_misp_timestamp );
+
+    // Add the frame timestamp to the metadata
+    kwiver::vital::timestamp frame_timestamp;
+    frame_timestamp.set_frame( d_frame_number );
+    if( d_have_frame_time )
+    {
+      frame_timestamp.set_time_usec( d_frame_time );
+    }
+    result.set_timestamp( frame_timestamp );
+
+    result.add< kwiver::vital::VITAL_META_VIDEO_URI >( video_path );
+
+    return { std::make_shared< kwiver::vital::metadata >( result ) };
   }
 
   // -------------------------------------------------------------------------------------
@@ -256,13 +243,10 @@ public:
     {
       retval = misp_time();
     }
-    else if ( time_source == "klv0601" )
+    else if ( time_source == "klv0601" || time_source == "klv0104" ||
+              time_source == "klv" )
     {
-      retval = klv_time( kwiver::arrows::klv::convert_metadata::MISB_0601 );
-    }
-    else if ( time_source == "klv0104" )
-    {
-      retval = klv_time( kwiver::arrows::klv::convert_metadata::MISB_0104 );
+      retval = klv_time();
     }
     else if ( time_source == "start_at_0" )
     {
@@ -362,7 +346,7 @@ public:
   } // current_time
 
   // -------------------------------------------------------------------------------------
-  bool klv_time( std::string type )
+  bool klv_time()
   {
     using namespace kwiver::vital;
 
@@ -378,32 +362,26 @@ public:
       }
 
       //It might be more accurate to get the second unique timestamp instead of the first
-      std::deque< vxl_byte > curr_md = d_video_stream.current_metadata();
-      auto klv_metadata = process_metadata( curr_md );
+      auto curr_md = d_video_stream.current_metadata();
+      auto klv_metadata =
+        process_metadata( { curr_md.cbegin(), curr_md.cend() } );
       if ( klv_metadata.size() > 0 )
       {
         // A metadata collection was created
         // check to see if it is of the desired type.
         for( auto meta : klv_metadata )
         {
-          // Test to see if the collection is from the specified standard (0104/0601)
-          if ( auto& origin = meta->find( VITAL_META_METADATA_ORIGIN ) )
+          if ( auto& ts = meta->find( VITAL_META_UNIX_TIMESTAMP ) )
           {
-            if (type == origin.as_string())
-            {
-              if ( auto& ts = meta->find( VITAL_META_UNIX_TIMESTAMP ) )
-              {
-                // Get unix timestamp as usec
-                meta_ts = static_cast< time_usec_t >( ts.as_uint64() );
+            // Get unix timestamp as usec
+            meta_ts = static_cast< time_usec_t >( ts.as_uint64() );
 
-                LOG_DEBUG( this->d_logger, "Found initial " << type <<
-                           " timestamp: " << meta_ts );
+            LOG_DEBUG( this->d_logger,
+                       "Found initial KLV timestamp: " << meta_ts );
 
-                d_have_abs_frame_time = true;
-                retval = true;
-              } // has time element
-            } // correct metadata type
-          } // has metadata origin
+            d_have_abs_frame_time = true;
+            retval = true;
+          } // has time element
         } // foreach over all metadata packets
       } // end if processed metadata collection
     }
@@ -422,7 +400,7 @@ public:
         c_use_metadata)
     {
       auto curr_md = d_video_stream.current_metadata();
-      auto metadata = process_metadata( curr_md );
+      auto metadata = process_metadata( { curr_md.cbegin(), curr_md.cend() } );
       if ( metadata.size() > 0 )
       {
         std::pair<vital::timestamp::frame_t, vital::metadata_vector>
@@ -947,7 +925,8 @@ vidl_ffmpeg_video_input
 
   // TODO: consider getting metadata from metadata map if it is present
   //       caching it there if not.
-  return d->process_metadata( d->d_video_stream.current_metadata() );
+  auto const curr_md = d->d_video_stream.current_metadata();
+  return d->process_metadata( { curr_md.cbegin(), curr_md.cend() } );
 }
 
 kwiver::vital::metadata_map_sptr
